requirements.txt
tangelo==0.9.0
github3.py==1.0.0a2
